First Backup
Display of the Sun, Earth, and Moon with the zodiac constellations. Earth rotation on its axis, revolution around the sun, and moon revolution around the earth is based on degrees of angle within time period. earth rotates 360 degrees in a day. earth revolves around the sun 1 degrees in a year for 360 days in a year. The moon rotates around the moon 360 degrees in 30 days. UI shows count of days and has buttons to move to the vernal and autumnal equinoxes.